+++ src/components/index.ts
@@ -2,8 +2,8 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './atoms/SvgIcon';
 export * from './cards/ExpandCardContent';
-export * from './icons/BicstoneIcon';
 export * from './layouts/CookieAlert';
 export * from './layouts/Footer';
 export * from './layouts/Head';
